feat(ui): rename privacy policy components with kebab-case & update lists rendering
- update components' folders to match style
- change rendering of 'points' inside components with new markup
