+++ app/(logged_in)/privacy-policy/page.tsx
@@ -4,16 +4,16 @@ import { Box } from '@mui/material';
 import { PAGE_IDS } from '~/constants/pageBlocks';
 import { Header } from '~/shared/components/header/Header';
 import { ContactUs } from '~/shared/components/privacy-policy/contact-us/ContactUs';
-import { Cookies } from '~/shared/components/privacy-policy/Cookies/Cookies';
 import { DataRetention } from '~/shared/components/privacy-policy/data-retention/DataRetention';
-import { DataUsage } from '~/shared/components/privacy-policy/DataUsage/DataUsage';
-import { DataWeCollect } from '~/shared/components/privacy-policy/DataWeCollect/DataWeCollect';
-import { GoogleAuth } from '~/shared/components/privacy-policy/GoogleAuth/GoogleAuth';
+import { DataUsage } from '~/shared/components/privacy-policy/data-usage/DataUsage';
+import { DataWeCollect } from '~/shared/components/privacy-policy/data-we-collect/DataWeCollect';
+import { GoogleAuth } from '~/shared/components/privacy-policy/google-auth/GoogleAuth';
 import { IntroSection } from '~/shared/components/privacy-policy/intro-section/IntroSection';
-import { NewsletterSubscription } from '~/shared/components/privacy-policy/NewsletterSubscription/NewsletterSubscription';
-import { SocialNetworks } from '~/shared/components/privacy-policy/SocialNetworks/SocialNetworks';
-import { TargetedAds } from '~/shared/components/privacy-policy/TargetedAds/TargetedAds';
+import { NewsletterSubscription } from '~/shared/components/privacy-policy/newsletter-subscription/NewsletterSubscription';
+import { SocialNetworks } from '~/shared/components/privacy-policy/social-networks/SocialNetworks';
+import { TargetedAds } from '~/shared/components/privacy-policy/targeted-ads/TargetedAds';
 import { UserRights } from '~/shared/components/privacy-policy/user-rights/UserRights';
+import { Cookies } from '~/shared/components/privacy-policy/сookies/Cookies';
 import { usePageEditor } from '~/shared/hooks/use-page-editor/usePageEditor';
 import { useSavePageBlocks } from '~/shared/hooks/use-save-page/UseSavePage';
 import { useStore } from '~/store';
