+++ app/shared/components/privacy-policy/social-networks/SocialNetworks.tsx
