+++ app/shared/components/privacy-policy/user-rights/UserRights.tsx
@@ -1,7 +1,7 @@
-import { Skeleton } from '@mui/material';
+import { Skeleton, Typography } from '@mui/material';
 import { JSONContent } from '@tiptap/react';
 
-import { EditableSectionList } from '../../accordion-blocks/editable-section-list/EditableSectionList';
+import ConfigurableList from '../../configurable-list/ConfigurableList';
 import CollapsibleBlock from '../../design-system/collapsible-block/CollapsibleBlock';
 import { CustomTextField } from '../../design-system/text-field/TextField';
 import { BLOCK_IDS, PAGE_IDS } from '~/constants/pageBlocks';
@@ -42,15 +42,30 @@ export const UserRights = () => {
         value={block.description[currentLocale]}
         onChange={(value) => handleChangeTitleText(value)}
       />
-      <EditableSectionList
-        title={block.title[currentLocale]}
-        onTitleChange={()=>{}}
-        items={points}
-        onChangeItem={()=>{}}
-        onCreateItem={()=>{}}
-        onDeleteItem={()=>{}}
-        sectionLabel="Пункти секції:"
-      />
+      {points.length > 0 && (
+        <>
+          <Typography variant="subtitle1" component="h4">
+            Пункти:
+          </Typography>
+          <ConfigurableList
+            items={points}
+            addBtnLabel="Додати пункт"
+            editable
+            onChange={() => { }}
+            onDelete={() => { }}
+            onCreate={() => { }}
+            renderItem={({ item, onChange }) => (
+              <CustomTextField
+                fieldType="formatting"
+                label="Текст пункту"
+                value={item.value}
+                onChange={(value) => onChange({ ...item, value })}
+              />
+            )}
+            separator={false}
+          />
+        </>
+      )}
       <CustomTextField
         fieldType="formatting"
         title="Додаткова інформація"
