+++ app/shared/components/privacy-policy/сookies/Cookies.tsx
@@ -1,35 +1,34 @@
 'use client';
 
-import { Skeleton } from '@mui/material';
+import { Skeleton, Typography } from '@mui/material';
 import { JSONContent } from '@tiptap/react';
 
-import { EditableSectionList, SectionListItem } from '../../accordion-blocks/editable-section-list/EditableSectionList';
+import ConfigurableList from '../../configurable-list/ConfigurableList';
 import CollapsibleBlock from '../../design-system/collapsible-block/CollapsibleBlock';
 import { CustomTextField } from '../../design-system/text-field/TextField';
 import { BLOCK_IDS, PAGE_IDS } from '~/constants/pageBlocks';
 import { ensureIds } from '~/lib/utils/ensureIds';
 import { usePageBlock } from '~/shared/hooks/use-page-block/usePageBlock';
 import { useStore } from '~/store';
-import { ItemWithId } from '~/types/store/pages/privacy-policy';
+import { ConfigurableListItem } from '~/types/accordionBlocks';
 
+export type CookiesPoint = ConfigurableListItem & { value: JSONContent };
 
-export const DataWeCollect = () => {
+export const Cookies = () => {
   const pageId = PAGE_IDS.PRIVACY_POLICY;
-  const blockId = BLOCK_IDS.DATA_WE_COLLECT;
+  const blockId = BLOCK_IDS.COOKIES;
 
   const { block } = usePageBlock(pageId, blockId);
   const currentLocale = useStore((state) => state.locale);
   const setField = useStore((state) => state.setField);
 
-
   if (!block) return <Skeleton sx={{ height: '60px' }} />;
 
-  const sectionsList: ItemWithId[] = ensureIds(block.sections);
+  const list = ensureIds(block.list);
 
-  const sections: SectionListItem[] = sectionsList.map((item) => ({
+  const cookiesPoints = list.map((item) => ({
     id: item.id,
-    title: item.subtitle[currentLocale],
-    description: item.list
+    value: item[currentLocale]
   }));
 
   const handleChangeTitleText = (value: JSONContent) => {
@@ -39,9 +38,8 @@ export const DataWeCollect = () => {
     });
   };
 
-
   return (
-    <CollapsibleBlock title="Які дані ми збираємо та чому">
+    <CollapsibleBlock title="Які cookie ми використовуємо">
       <CustomTextField
         fieldType="formatting"
         title="Вступний текст секції"
@@ -49,22 +47,38 @@ export const DataWeCollect = () => {
         onChange={(value) => handleChangeTitleText(value)}
       />
 
-      <EditableSectionList
-        title={block.title[currentLocale]}
-        onTitleChange={() => { }}
-        items={sections}
-        onChangeItem={() => { }}
-        onCreateItem={() => { }}
-        onDeleteItem={() => { }}
-        sectionLabel="Пункти секції:"
-      />
-
+      {cookiesPoints.length > 0 && (
+        <>
+          <Typography variant="subtitle1" component="h4">
+             Пункти:
+          </Typography>
+          <ConfigurableList<CookiesPoint>
+            items={cookiesPoints}
+            addBtnLabel="Додати пункт"
+            editable
+            onChange={()=>{}}
+            onDelete={()=>{}}
+            onCreate={()=>{}}
+            renderItem={({ item, onChange }) => (
+              <CustomTextField
+                fieldType="formatting"
+                label="Текст пункту"
+                value={item.value}
+                onChange={(value) => onChange({ ...item, value })}
+              />
+            )}
+            separator={false}
+          />
+        </>
+      )}
+ 
       <CustomTextField
         fieldType="formatting"
         title="Додаткова інформація"
         value={block.note[currentLocale]}
         onChange={(value) => handleChangeTitleText(value)}
       />
+
     </CollapsibleBlock>
   );
 };
